% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cariboutools-package.R
\docType{package}
\name{cariboutools-package}
\alias{cariboutools}
\alias{cariboutools-package}
\title{cariboutools: Helpers for the YT_caribou Modelling Workflow}
\description{
Helper functions for the Yukon caribou / anthropogenic-disturbance modelling workflow: building the study area and template rasters, and other shared utilities used across the prep-fit and predict targets pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FOR-CAST/cariboutools}
  \item Report bugs at \url{https://github.com/FOR-CAST/cariboutools/issues}
}

}
\author{
\strong{Maintainer}: Alex M. Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Alex M. Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
}

}
\keyword{internal}
